requirements.txt
streamlit==1.39.0
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
PyYAML==6.0.2
python-dateutil==2.9.0.post0
